refactor(tauri/ui): extract InputVaultId into InputHex
